Adjust m7_o1 video editor validation and description
